requirements.txt
flask
cfbd
firebase-admin
dataclasses-json
python-dotenv
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib